240311_혜인_혜인+다선

+++ Final_StudyWithHour/src/com/test/mvc/GuestDeleteController.java
@@ -17,32 +17,41 @@ public class GuestDeleteController implements Controller
 	{
 		// 액션 코드
 		ModelAndView mav = new ModelAndView();
-		HttpSession session = request.getSession();
 		
 		// 세션 처리과정
+		// 게스트 코드가 있는 경우에만 접근 가능
+		HttpSession session = request.getSession();
 		
+		if (session.getAttribute("guCode")==null)
+		{
+			// 게스트 코드가 없는 경우 로그인 폼으로 이동
+			mav.setViewName("redirect:loginform.do");
+			return mav;
+		}
+				
 		
 		String guCode = request.getParameter("guCode");
 		String reasonCode = request.getParameter("reasonCode");
 		
+		GuestDAO dao = new GuestDAO();
 		try
 		{
-			GuestDAO dao = new GuestDAO();
+			dao.connection();
 			
 			dao.remove(guCode, reasonCode);
 			
 			session.removeAttribute("guCode");
 			mav.setViewName("MainPage.jsp");
 			
-			dao.close();
-			
 		}
 		catch (Exception e) 
 		{
 			System.out.println(e.toString());
 		}
-		
-			
+		finally 
+		{
+			dao.close();
+		}
 		
 		return mav;
